backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
datasets==2.14.5
transformers==4.33.2
sentence-transformers==2.2.2
pypdf2==3.0.1
weaviate-client==3.24.1
openai==0.28.1
langchain==0.0.285
python-multipart==0.0.6
python-dotenv==1.0.0